requirements.txt
paho-mqtt~=2.1.0
Flask~=3.0.3
requests~=2.32.3
pycryptodome~=3.20.0